+++ Source/Game/Play/Item/ShowItemWarningColorBlock.c
@@ -24,7 +24,7 @@ void ShowItemWarningColorBlock(Player* player) {
 
 static void UpdateEntityItemWarningColorBlock(Entity* entity) {
 	ENTITY_INST_DATA_PTR(BasicEntityInstanceData, data, entity);
-	DisplayObject(&data->objectTable[entity->warningFrames], data->y, data->x, data->palNum, 124u);
+	DisplayObject(&data->objectTable[entity->warningFrames], data->y, data->x, (uint8_t)data->palNum, 124u);
 
 	if (CurrentPauseMode < PAUSEMODE_GAME && entity->warningFrames >= 64) {
 		FreeEntity(entity);
