+++ Source/Game/Play/Item/ShowItemWarningRollRoll.c
@@ -24,7 +24,7 @@ void ShowItemWarningRollRoll(Player* player) {
 
 static void UpdateEntityItemWarningRollRoll(Entity* entity) {
 	ENTITY_INST_DATA_PTR(BasicEntityInstanceData, data, entity);
-	DisplayObject(&data->objectTable[entity->warningFrames], data->y, data->x, data->palNum, 124u);
+	DisplayObject(&data->objectTable[entity->warningFrames], data->y, data->x, (uint8_t)data->palNum, 124u);
 	if (CurrentPauseMode < PAUSEMODE_GAME && ++entity->warningFrames >= 64) {
 		FreeEntity(entity);
 	}
