+++ Source/Game/Play/Item/ShowItemWarningShotgun.c
@@ -25,7 +25,7 @@ void ShowItemWarningShotgun(Player* player) {
 
 static void UpdateEntityItemWarningShotgun(Entity* entity) {
 	ENTITY_INST_DATA_PTR(BasicEntityInstanceData, data, entity);
-	DisplayObject(&data->objectTable[entity->warningFrames], data->y, data->x, data->palNum, 124u);
+	DisplayObject(&data->objectTable[entity->warningFrames], data->y, data->x, (uint8_t)data->palNum, 124u);
 
 	if (CurrentPauseMode < PAUSEMODE_GAME && ++entity->warningFrames >= 64) {
 		FreeEntity(entity);
